requirements.txt
# Core scientific computing
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
plotly>=5.14.0
scipy>=1.10.0

# Computer vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0

# Deep learning frameworks
torch>=2.0.0
torchvision>=0.15.0

# Real detection models
ultralytics>=8.0.0  # YOLOv8

# Segmentation (SAM - Segment Anything Model)
segment-anything>=1.0

# 3D simulation
pybullet>=3.2.5

# Path planning
scikit-learn>=1.3.0  # For KD-tree in RRT*
networkx>=3.0  # For graph-based planning

# Utilities
pyyaml>=6.0
Pillow>=10.0.0
tqdm>=4.65.0

# Optional but recommended
# pytest>=7.3.0  # For testing
